requirements.txt
fpdf==1.7.2
fpdf2==2.8.2
PyYAML==6.0.2
robotframework==7.2
robotframework-seleniumlibrary==6.0.0
robotframework-yamllibrary==1.0.4